Desert update Fixed - Preventing stuff from occuring if random event or agility. - Tanner added back. - Barbarian agility fixed on the wall.

+++ game-server/src/main/java/net/dodian/uber/game/model/player/packets/incoming/DropItem.java
@@ -14,6 +14,9 @@ public class DropItem implements Packet {
         client.getInputStream().readUnsignedByte();
         client.getInputStream().readUnsignedByte();
         int slot = client.getInputStream().readUnsignedWordA();
+        if (client.randomed || client.UsingAgility) {
+            return;
+        }
         if (droppedItem == 5733) {
             client.deleteItem(droppedItem, slot, 1);
             client.send(new SendMessage("A magical force removed this item from your inventory!"));
